src/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Migrate>=3.0.0
python-dotenv>=0.19.0
psycopg2-binary # If using PostgreSQL, otherwise optional
